personal_website/requirements.txt
Django==1.9.12
PyGithub==1.35
PyJWT==1.5.3
